Move dim utilities to `dim.c`
